Grammar/spelling/flow cleanup, add short intro paragraph
This cleans up grammar, spelling, and overall flow of the article a tiny
bit. I also added a short intro paragraph, since it helps with the flow
a bit more, and I thought it was appropriate to be somewhat celebratory
(if you can get over some of my lame jokes/references).